Split core files
