+++ src/pages/GameSuccessPage.tsx
@@ -1,7 +1,8 @@
 import { Navigate } from "react-router-dom";
 import styled from "@emotion/styled";
 
-import { useGameConfig, useGameResult } from "core/game";
+import { useGameConfig } from "core/gameConfig";
+import { useGameResult } from "core/gameResult";
 import PageWrapper from "pages/PageWrapper";
 import { paths } from "pages/PageRouter";
 import NewGameButton from "components/NewGame/NewGameButton";
